Rename some html elements, repair css styles for buttons
