+++ src/main/webapp/reactjs/src/css/header-and-body.css
@@ -96,24 +96,24 @@ button:focus, textarea {
     margin-top: 15px;
 }
 
-.item-search-field {
-    padding: 1em;
-    border-radius: 6px;
-    border: none;
-    font-size: 15px;
-    height: 36px;
-    width: 20vw;
-    min-width: 200px;
-}
+/*.item-search-field {*/
+/*    padding: 1em;*/
+/*    border-radius: 6px;*/
+/*    border: none;*/
+/*    font-size: 15px;*/
+/*    height: 36px;*/
+/*    width: 20vw;*/
+/*    min-width: 200px;*/
+/*}*/
 
-.distance-inputs {
-    padding: 1em;
-    border-radius: 6px;
-    border: none;
-    font-size: 15px;
-    height: 36px;
-    width: 150px;
-}
+/*.distance-inputs {*/
+/*    padding: 1em;*/
+/*    border-radius: 6px;*/
+/*    border: none;*/
+/*    font-size: 15px;*/
+/*    height: 36px;*/
+/*    width: 150px;*/
+/*}*/
 
 .sorting-option-inputs-list {
     float: right;
@@ -179,7 +179,7 @@ button:focus, textarea {
     padding: 20px;
 }
 
-.popup-content button {
+.recoveryPassBtn {
     background: none;
     border: none;
     text-align: right;
@@ -191,6 +191,8 @@ button:focus, textarea {
 .popup-content button:hover{
     cursor: pointer;
 }
+
+
 
 .login-data-input {
     border-radius: 10px;
@@ -266,11 +268,9 @@ button:focus, textarea {
     align-items: end;
 }
 
-
 .button12 {
     color: white;
     place-self: center;
-    text-align: center;
     font-size: inherit;
     background: none;
     border-width: 1px;
@@ -281,9 +281,7 @@ button:focus, textarea {
     height: 45px;
     padding: 0;
     font-weight: 900;
-    padding-top: 20px !important;
 }
-
 .button12:hover {
     background-color: rgba(255, 255, 255, 0.6);
     transition-duration: 0.5s;
@@ -292,20 +290,6 @@ button:focus, textarea {
 
 a {
     text-decoration: none;
-}
-
-
-.button1 {
-    color: white;
-    place-self: center;
-    text-align: center;
-    letter-spacing: 3px;
-    border-width: 1px;
-    border-radius: 10px;
-    border-style: solid;
-    width: 120px;
-    margin-top: 10px;
-    padding: 12px;
 }
 
 h5 a {
@@ -324,6 +308,23 @@ h5 a:hover {
     background-color: rgba(255, 255, 255, 0.6);
     transition-duration: 0.5s;
     cursor: pointer;
+}
+
+.loginRecBtn{
+    color:white;
+    place-self: center;
+    text-align:center;
+    font-size: inherit;
+    background: none;
+    border-width: 1px;
+    border-radius: 10px;
+    border-style: solid;
+    border-color: white;
+    width: 250px;
+    height:45px;
+    padding:0;
+    font-weight: 900;
+    vertical-align: center;
 }
