+++ src/main/webapp/reactjs/src/css/index.css
@@ -92,23 +92,6 @@
     width: 1.5vw;
 }
 
-.loginBtn{
-    color:white;
-    place-self: center;
-    text-align:center;
-    font-size: inherit;
-    background: none;
-    border-width: 1px;
-    border-radius: 10px;
-    border-style: solid;
-    border-color: white;
-    width: 250px;
-    height:45px;
-    padding:0;
-    font-weight: 900;
-    padding-top: 20px !important;
-}
-
 #categories-area{
     position: absolute;
     top: 100vh;
